class 파일 분할
