+++ t1.txt
@@ -1,18 +1,14 @@
-[1, 'x', 0, 1, 2]
-[0, 'x', 241, 1, 2]
-[1, 'z', 388, 1, 2]
-[0, 'z', 737, 1, 2]
-[1, 'z', 1907, 1, 2]
-[0, 'z', 2309, 1, 2]
-[1, 'x', 2411, 1, 2]
-[0, 'x', 2688, 1, 2]
-[1, 'c', 2793, 1, 2]
-[0, 'c', 2961, 1, 2]
-[1, 'v', 3191, 1, 2]
-[0, 'v', 3461, 1, 2]
-[1, 'b', 3535, 1, 2]
-[0, 'b', 3777, 1, 2]
-[1, 'n', 3899, 1, 2]
-[0, 'n', 4087, 1, 2]
-[1, 'm', 4233, 1, 2]
-[0, 'm', 4562, 1, 2]
+[1, 'c', 0, 1, 2]
+[0, 'c', 396, 1, 2]
+[1, 'x', 404, 1, 2]
+[0, 'x', 775, 1, 2]
+[1, 'z', 802, 1, 2]
+[0, 'z', 1151, 1, 2]
+[1, 'x', 1225, 1, 2]
+[0, 'x', 1498, 1, 2]
+[1, 'c', 1623, 1, 2]
+[0, 'c', 1900, 1, 2]
+[1, 'c', 2071, 1, 2]
+[0, 'c', 2321, 1, 2]
+[1, 'c', 2490, 1, 2]
+[0, 'c', 2849, 1, 2]
